initial support for sn74hc595 shift register
